+++ R/matrixData.R
@@ -1,22 +1,21 @@
 #' Check perseus compatibility of an object
 #'
+#' @rdname MatrixDataCheck
+#'
 #' @title MatrixDataCheck: a function to check the validity of an object as a perseus data frame
 #'
 #' @param object object to check consistency with perseus data frames
 #' @param ... additional arguments passed to the respective method
 #' @param main Main Data frame
 #' @param annotationRows Rows containing annotation information
-#' @param annotationCols Collumns containing annotation information
+#' @param annotationCols Columns containing annotation information
 #' @param descriptions Descriptions of all the columns
 #' @param imputeData Is imputed or not
 #' @param qualityData quality number
 #' @param all_colnames The colnames to be used
 #'
-#'
 #' @return a logical indicating the validity of the object
 #' (or series of objects) as a perseus DF or the string of errors
-#'
-#' @rdname MatrixDataCheck
 #'
 #' @export
 #'
@@ -30,7 +29,6 @@
 #'     annotRows=data.frame(x=factor(c('1','1'))))
 #'
 #' MatrixDataCheck(mat)
-#'
 #'
 MatrixDataCheck <- function(object, ...) {
   UseMethod("MatrixDataCheck", object)
@@ -280,7 +278,7 @@ names.matrixData <- getNames
 
 #' Get main columns
 #'
-#' Gets the main collumns (main matrix) of a \code{\link[PerseusR]{matrixData}}
+#' Gets the main columns (main matrix) of a \code{\link[PerseusR]{matrixData}}
 #' object as a data.frame object
 #'
 #' @param mdata matrixData